Financial time series for Tsay (2005, ch. 1)

Financial time series used in examples in chapter 1.

data(d.ibmvwewsp6203)
data(d.intc7303)
data(d.3m6203)
data(d.msft8603)
data(d.c8603)
data(m.ibmvwewsp2603)
data(m.intc7303)
data(m.3m4603)
data(m.msft8603)
data(m.c8603)
data(m.gs10)
data(m.gs1)
data(d.fxjp00)
data(m.fama.bond5203)
data(m.gs3)
data(m.gs5)
data(w.tb3ms)
data(w.tb6ms)

Format

Objects of class zoo giving simple returns for each trading period (day, week or month) for different periods, with different start dates but typically running to the end of 2003.

d.ibmvwewsp6203, m.ibmvwewsp2603 Zoo objects with 4 columns (IBM, VW, EW, and SP). Daily data starts with 1962-07-03. Monthly data starts with 1926-01-30.

d.intc7303, m.intc7303 Matrices of class zoo with a single column "Intel" starting from January 1973.

d.3m6203, m.3m6203 Matrices of class zoo with a single column "MMM". Daily data starts with 1962-07-03. Monthly data starts with 1946-02-28.

d.msft8603, m.msft8603 Matrices of class zoo with a single column "MSFT" starting from 1906-03-14.

d.c8603, m.c8603 Matrix of class zoo with a single column "C" starting from 1986-10-30.

m.gs10, m.gs1 Monthly 10-yr and 1-yr Treasury constant maturity rates (4/53-3/04)

d.fxjp00 Daily exchange rate between U.S. dollar and Japanese yen

m.fama.bond5203 Monthly bond returns as follows:

m1.121-12m

m24.3624-36m

m48.6048-60m

m61.12061-120m

m.gs3, m.gs5 Monthly 3-yr and 5-yr Treasury constant maturity rates

w.tb3ms, w.tb6ms Weekly Treasury Bill rates

Details

The first 16 of these objects contain daily and monthly simple returns for 8 financial time series analyzed Tsay (2005, Table1.2). These 8 are SP (Standard & Poors), EW, IBM, Intel, Microsoft, and Citi-Group, beginning at different times and running to the end of 2003.

The others are used elsewhere in chapter 1.

Source

https://faculty.chicagobooth.edu/ruey-s-tsay/teaching

References

Ruey Tsay (2005) Analysis of Financial Time Series, 2nd ed. (Wiley, ch. 1)

See also

Examples

# First half of Table 1.2:
data(d.ibmvwewsp6203)
data(d.intc7303)
data(d.3m6203)
data(d.msft8603)
data(d.c8603)
(Daily.Simple.Returns.pct <- rbind(
    SP = FinTS.stats(100*d.ibmvwewsp6203[, "SP"]),
    VW = FinTS.stats(100*d.ibmvwewsp6203[, "VW"]),
    EW = FinTS.stats(100*d.ibmvwewsp6203[, "EW"]),
    IBM= FinTS.stats(100*d.ibmvwewsp6203[, "IBM"]),
    Intel=FinTS.stats(100*d.intc7303[, "Intel"]),
    MMM= FinTS.stats(100*d.3m6203[, "MMM"]),
    MSFT=FinTS.stats(100*d.msft8603[, 'MSFT']),
    C  = FinTS.stats(100*d.c8603[, "C"])
) )
#>            Start  Size       Mean Standard.Deviation    Skewness
#> SP    1962-07-03 10446 0.03312081          0.9453075 -0.94499859
#> VW    1962-07-03 10446 0.04538579          0.8912131 -0.76008356
#> EW    1962-07-03 10446 0.08502202          0.7255660 -0.88633597
#> IBM   1962-07-03 10446 0.05233745          1.6481052  0.07749371
#> Intel 1973-01-02  7828 0.13073441          2.9983212 -0.15828297
#> MMM   1962-07-03 10446 0.05430031          1.4644984 -0.28348680
#> MSFT  1986-03-14  4493 0.15738059          2.5048308 -0.24946227
#> C     1986-10-30  4333 0.11003623          2.2891492  0.10079641
#>       Excess.Kurtosis Minimum Maximum
#> SP          25.764589 -20.470   9.100
#> VW          18.321499 -17.140   8.660
#> EW          13.419351 -10.390   6.950
#> IBM         10.214412 -22.963  13.164
#> Intel        5.849641 -29.572  26.378
#> MMM         12.868398 -25.979  11.538
#> MSFT         8.749528 -30.116  19.565
#> C            6.787964 -21.739  20.755

(Daily.log.Returns.pct <- rbind(
    SP = FinTS.stats(100*log(1+d.ibmvwewsp6203[, "SP"])),
    VW = FinTS.stats(100*log(1+d.ibmvwewsp6203[, "VW"])),
    EW = FinTS.stats(100*log(1+d.ibmvwewsp6203[, "EW"])),
    IBM= FinTS.stats(100*log(1+d.ibmvwewsp6203[, "IBM"])),
    Intel=FinTS.stats(100*log(1+d.intc7303[,"Intel"])),
    MMM= FinTS.stats(100*log(1+d.3m6203[, "MMM"])),
    MSFT=FinTS.stats(100*log(1+d.msft8603[, 'MSFT'])),
    C  = FinTS.stats(100*log(1+d.c8603[, "C"]))
) )
#>            Start  Size       Mean Standard.Deviation   Skewness Excess.Kurtosis
#> SP    1962-07-03 10446 0.02861753          0.9504892 -1.4058397       36.907587
#> VW    1962-07-03 10446 0.04138657          0.8945797 -1.0620475       23.906167
#> EW    1962-07-03 10446 0.08234596          0.7275677 -1.0519055       14.703639
#> IBM   1962-07-03 10446 0.03874377          1.6490332 -0.2548118       12.602898
#> Intel 1973-01-02  7828 0.08549078          3.0128747 -0.5442402        7.536558
#> MMM   1962-07-03 10446 0.04352374          1.4693908 -0.6883163       20.055791
#> MSFT  1986-03-14  4493 0.12573133          2.5181738 -0.7307891       13.225870
#> C     1986-10-30  4333 0.08381041          2.2893291 -0.2109124        7.467405
#>         Minimum   Maximum
#> SP    -22.90359  8.709471
#> VW    -18.80177  8.305356
#> EW    -10.97033  6.719125
#> IBM   -26.08844 12.366791
#> Intel -35.05793 23.410723
#> MMM   -30.08213 10.919515
#> MSFT  -35.83335 17.868997
#> C     -24.51208 18.859351