List of the names of files downloaded from the "Analysis of Financial Data" web site.

A list organized by chapter and text vs. exercises of the files downloaded from the web site associated with Tsay (2005) Analysis of Financial Time Series, 2nd ed. (Springer) and stored in "~library/FinTS/scripts/TsayFiles". These facilitate the process of creating and updating the 'FinTS' package (and documenting the creation process).

Usage

TsayFiles

Format

TsayFiles A list with names 'ch01', 'ch02', ..., 'ch12' for components describing the files associated with the corresponding chapter. Each chapter component is a list with 'text' and 'exercises' components, where 'text' and 'exercises' are each a character array giving names for 'data', 'file', 'url', and 'found' for the data referenced in the text or exercises of that chapter:

data 'file' without the extension, e.g, 'd-ibmvwewsp6203' for daily simple returns of IBM, VW, EW, SP (7/3/62-12/31/03) = 'file' without the extension.

file short file name = 'data' plus the extension, e.g., 'd-ibmvwewsp6203'.txt' for daily simple returns of IBM, VW, EW, SP (7/3/62-12/31/03)

url universal resource locator for the data, e.g., "http://faculty.chicagogsb.edu/ruey.tsay/teaching/fts2/d-ibmvwewsp6203.txt"

found 'TRUE' if the data were found, 'FALSE' if the attempt to access the url failed.

NOTES: (1) 13 files are referenced twice, and 2 are referenced three times on the web page. This redundancy is retained in 'TsayFiles'. (2) A few files (most noticably some with with '.dat' extension) are referenced in the HTML code without an apparent visible link. These 'invisible files' are retained in 'TsayFiles'.

FinTS.url A character string giving the universal resource locator (URL) associated with the Tsay (2005) book: FinTS.url <- "https://faculty.chicagobooth.edu/ruey-s-tsay/teaching"

Source

Tsay (2005) Analysis of Financial Time Series, 2nd ed. (Springer)

Examples

data(TsayFiles)
TsayFiles$ch01$exercises
#>                  data           file              
#> d-3stock.txt     "d-3stock"     "d-3stock.txt"    
#> d-tyc9099.dat    "d-tyc9099"    "d-tyc9099.dat"   
#> m-ibm3dx7503.txt "m-ibm3dx7503" "m-ibm3dx7503.txt"
#> m-mel7399.dat    "m-mel7399"    "m-mel7399.dat"   
#> d-fxca00.txt     "d-fxca00"     "d-fxca00.txt"    
#> d-fxuk00.txt     "d-fxuk00"     "d-fxuk00.txt"    
#> d-fxjp00.txt     "d-fxjp00"     "d-fxjp00.txt"    
#> d-fxeu00.txt     "d-fxeu00"     "d-fxeu00.txt"    
#>                  url                                                                     
#> d-3stock.txt     "http://faculty.chicagogsb.edu/ruey.tsay/teaching/fts2/d-3stock.txt"    
#> d-tyc9099.dat    "http://faculty.chicagogsb.edu/ruey.tsay/teaching/fts2/d-tyc9099.dat"   
#> m-ibm3dx7503.txt "http://faculty.chicagogsb.edu/ruey.tsay/teaching/fts2/m-ibm3dx7503.txt"
#> m-mel7399.dat    "http://faculty.chicagogsb.edu/ruey.tsay/teaching/fts2/m-mel7399.dat"   
#> d-fxca00.txt     "http://faculty.chicagogsb.edu/ruey.tsay/teaching/fts2/d-fxca00.txt"    
#> d-fxuk00.txt     "http://faculty.chicagogsb.edu/ruey.tsay/teaching/fts2/d-fxuk00.txt"    
#> d-fxjp00.txt     "http://faculty.chicagogsb.edu/ruey.tsay/teaching/fts2/d-fxjp00.txt"    
#> d-fxeu00.txt     "http://faculty.chicagogsb.edu/ruey.tsay/teaching/fts2/d-fxeu00.txt"    
#>                  found  
#> d-3stock.txt     "TRUE" 
#> d-tyc9099.dat    "FALSE"
#> m-ibm3dx7503.txt "TRUE" 
#> m-mel7399.dat    "FALSE"
#> d-fxca00.txt     "TRUE" 
#> d-fxuk00.txt     "TRUE" 
#> d-fxjp00.txt     "TRUE" 
#> d-fxeu00.txt     "TRUE"